requirements.txt
django
django-cors-headers
djangorestframework
drf-yasg
whitenoise
gunicorn
numpy
sklearn